ADD lauchn game socket + current word first player + display each player per round + fix bug order player bdd
