+++ mot_de_passe/frontend/src/style/index.scss
@@ -3,7 +3,7 @@
 @import "./pages/_loginPlayers.scss";
 @import "./pages/_loginGameMaster.scss";
 @import "./pages/_home.scss";
-@import "./pages/_gamePlayers.scss";
+@import "./pages/_gamePlayersSpeaker.scss";
 @import "./pages/_gameGM.scss";
 @import "./pages/_gameViewers.scss";
 @import "./pages/_waitingroom.scss";
