+++ mot_de_passe/frontend/src/style/pages/_gamePlayersSpeaker.scss
